server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
openai==1.3.0
pydub==0.25.1
SpeechRecognition==3.10.0
requests==2.31.0
eventlet==0.33.3
python-socketio==5.8.0
whisper==1.1.10
numpy==1.24.3
ffmpeg-python==0.2.0
torch==2.6.0
